+++ fivetran/framework/core/model/quickstart_package.go
@@ -9,11 +9,18 @@ import (
 )
 
 type QuickstartPackage struct {
-	Id              	types.String `tfsdk:"id"`
-	Name         		types.String `tfsdk:"name"`
-	Version  			types.String `tfsdk:"version"`
-	ConnectorTypes  	types.Set 	 `tfsdk:"connector_types"`
-	OutputModelNames 	types.Set    `tfsdk:"output_model_names"`
+	Id               types.String `tfsdk:"id"`
+	Name             types.String `tfsdk:"name"`
+	Version          types.String `tfsdk:"version"`
+	ConnectorTypes   types.Set    `tfsdk:"connector_types"`
+	OutputModelNames types.Set    `tfsdk:"output_model_names"`
+	ConfigurableVars types.Map    `tfsdk:"configurable_vars"`
+}
+
+var configurableVarAttrTypes = map[string]attr.Type{
+	"type":           types.StringType,
+	"description":    types.StringType,
+	"allowed_values": types.ListType{ElemType: types.StringType},
 }
 
 func (d *QuickstartPackage) ReadFromResponse(ctx context.Context, resp transformations.QuickstartPackageResponse) {
@@ -28,8 +35,32 @@ func (d *QuickstartPackage) ReadFromResponse(ctx context.Context, resp transform
 	d.ConnectorTypes, _ = types.SetValue(types.StringType, connectors)
 
 	models := []attr.Value{}
-	for _, connector := range resp.Data.OutputModelNames {
-		models = append(models, types.StringValue(connector))
+	for _, model := range resp.Data.OutputModelNames {
+		models = append(models, types.StringValue(model))
 	}
 	d.OutputModelNames, _ = types.SetValue(types.StringType, models)
+
+	d.ConfigurableVars = configurableVarsFromResponse(resp.Data.ConfigurableVars)
+}
+
+func configurableVarsFromResponse(vars map[string]transformations.ConfigurableVarDefinition) types.Map {
+	if len(vars) == 0 {
+		return types.MapNull(types.ObjectType{AttrTypes: configurableVarAttrTypes})
+	}
+	items := map[string]attr.Value{}
+	for name, v := range vars {
+		allowedVals := []attr.Value{}
+		for _, av := range v.AllowedValues {
+			allowedVals = append(allowedVals, types.StringValue(av))
+		}
+		allowedList, _ := types.ListValue(types.StringType, allowedVals)
+		obj, _ := types.ObjectValue(configurableVarAttrTypes, map[string]attr.Value{
+			"type":           types.StringValue(v.Type),
+			"description":    types.StringValue(v.Description),
+			"allowed_values": allowedList,
+		})
+		items[name] = obj
+	}
+	result, _ := types.MapValue(types.ObjectType{AttrTypes: configurableVarAttrTypes}, items)
+	return result
 }